docs(filter-button): update stories
