+++ packages/components/src/FilterButton/docs/FilterButton.stories.tsx
@@ -8,7 +8,7 @@ export default {
   parameters: {
     docs: {
       description: {
-        component: "Trigger button for Filter.",
+        component: "To be used with Filter.",
       },
     },
   },
